adds missing css from last commit...
not sure how or why this happened but on the commit of fix_triangles on june 6, 2019, the css didn't get committed. Everything else did but for some reason the css didn't. Running git status showed the "untracked changes" as if i had never committed them. after adding them again while on the same branch of fix_triangles, the commit didn't work and also didn't throw an error. it waited for additional input but I know of no other input other than just hitting enter after the commit message. 
used control + c to get out of this. had to re commit the css which worked. but when pushing this, no pull request ever came up on github. luckily I copied the entire css on my local clipboard and have just pasted it in here. after committing this, will try a git pull and hopefully get the changes on my local.